requirements.txt
Django==6.0.4
Pillow==12.2.0
django-import-export==4.4.0
openpyxl==3.1.5
python-escpos==3.1     # Impresión ESC/POS en impresoras térmicas
pywin32==311           # Requerido por python-escpos en Windows (Win32Raw)

# Dependencias indirectas (generadas automáticamente)
asgiref==3.11.1
diff-match-patch==20241021
et_xmlfile==2.0.0
sqlparse==0.5.5
tablib==3.9.0
tzdata==2026.1